Fix createOrDestroyTilesIfNeeded repaint rect extension algorithm
Fix a regression induced with commit a45c892ec9dc6b63f5106ced40a4fb7d168c7aad. This
commit is basically ensuring that part of the screen that contains layer that are
removed are repainted. Unfortunatly the createOrDestroyTilesIfNeeded was to eager to
expand this rect, and in some cases it became bigger than the size of the backing
store itself, resulting in memory corruption.