Fix DFB_SUPPORT flag in wmph
The conditionnal directfb code in wymedia player helper is controlled using ifdef
definition. This works fine if the flag is unset/set, but fail when the flag is
set to 0 (which is expected to work as if it was not set). To fix this, use if
ENABLE(DFB_SUPPORT) instruction instead. In particular, this issue prevent webkit
to properly build when the --minimal flag is set.