Only compile accelerated compositing support in OpenGL builds
The transform of composited layers doesn't track which area of
the screen is dirty and thus needs a full repaint of the screen
when accelerated compositing mode is entered. On lower end hardware
this prevents achieving higher framerates and it's better in this
case to disable the creation of layers while trying to limit the
complexity of the layers on the HTML side.

This has the same effect as setting to false
QWebSettings::AcceleratedCompositingEnabled, disabling it at
compile time also saves a few hundreds of kbytes on the binary.