Inspector: Prevent a full-refresh of the webview
hideHighlight is called inconditionally each time that the inspector wants
to make sure that the highlight is visible, and does so in many situations.
We should avoid updating the complete view in this case to avoid hiding
the size of the area that that is really requested to be repainted when
taking timeline traces.